Move JS(X) files to TS(X)
