+++ packages/cubejs-client-react/src/hooks/is-mounted.ts
